\name{forestfires}
\alias{forestfires}
\docType{data}
\title{Forest Fires Data}
\description{
Meteorological and fire-weather measurements for forest fires in the
northeast region of Portugal.
}
\usage{data(forestfires)}
\format{
A data frame with 517 observations and 13 variables:
\describe{
  \item{X}{Horizontal spatial coordinate.}
  \item{Y}{Vertical spatial coordinate.}
  \item{month}{Month of the year.}
  \item{day}{Day of the week.}
  \item{FFMC}{Fine Fuel Moisture Code.}
  \item{DMC}{Duff Moisture Code.}
  \item{DC}{Drought Code.}
  \item{ISI}{Initial Spread Index.}
  \item{temp}{Temperature in degrees Celsius.}
  \item{RH}{Relative humidity in percent.}
  \item{wind}{Wind speed in kilometres per hour.}
  \item{rain}{Outside rain in millimetres per square metre.}
  \item{area}{Burned forest area in hectares.}
}
}
\source{
Cortez, P. and Morais, A. (2007), Forest Fires dataset, UCI Machine Learning
Repository, \url{https://archive.ics.uci.edu/dataset/162/forest+fires}.
}
\keyword{datasets}
